requirements.txt
griptape>=1.1.1
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
zoomus>=1.1.1
PyJWT>=2.0.0